# src/public/docs/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Earthid SSI Layer",
    "version": "2.0.0",
    "description": "This the SSI Layer of Earthid",
    "contact": {
      "email": "vishnuK@myearth.id",
      "name": "API Support"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Issuer Appnet - Dev"
    },
    {
      "url": "http://localhost:8081",
      "description": "Verifier Appnet - Dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    }
  },
  "paths": {
...
    "/api/issuer/issuance-request": {
      "post": {
        "summary": "Create a Schema for Verifiable Credential.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Issuer"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "givenName": {
                    "type": "string",
                    "example": "Vishnu",
                    "description": "first name"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "sandra",
                    "description": "last name"
                  },
                  "userName": {
                    "type": "string",
                    "example": "VishnuSkrishnan",
                    "description": "user name"
                  },
                  "membershipType": {
                    "type": "string",
                    "example": "fulltime",
                    "description": "type"
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "example": "22/05/1996",
                    "description": "DOB"
                  }
                },
                "required": [
                  "givenName",
                  "lastName"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Schema created",
            "content": {
              "image/png": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "example": "The Schema is created"
                    },
                    "status": {
                      "type": "number",
                      "example": 201
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Verifiable Credential creation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Schema creation failed!!"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/issuer/schema/{schemaName}": {
      "get": {
        "summary": "Retrieve schema Details",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Issuer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "schemaName",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Schema Details retrieval success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "$schema": {
                          "type": "string",
                          "format": "uri",
                          "example": "http://json-schema.org/draft-04/schema#"
                        },
                        "title": {
                          "type": "string",
                          "example": "MembershipType"
                        },
                        "description": {
                          "type": "string",
                          "example": "Membership Document for GBA Credential"
                        },
                        "type": {
                          "type": "string",
                          "example": "object"
                        },
                        "properties": {
                          "type": "object",
                          "properties": {
                            "firstName": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "pattern": {
                                  "type": "string",
                                  "example": "[a-zA-Z]+$"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "First name of the user"
                                },
                                "maxLength": {
                                  "type": "string",
                                  "example": "4"
                                }
                              }
                            },
                            "lastName": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "pattern": {
                                  "type": "string",
                                  "example": "[a-zA-Z]+$"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "Last name of the user"
                                },
                                "maxLength": {
                                  "type": "string",
                                  "example": "4"
                                }
                              }
                            }
                          }
                        },
                        "required": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "example": [
                            "firstName"
                          ]
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Schema is retrieved"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Schema details retrieval failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid Schema"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/issuer/schemas": {
      "get": {
        "summary": "Retrieve All Schemas of Issuer",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Issuer"
        ],
        "responses": {
          "200": {
            "description": "Issuer's Schemas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "description": "List of all schemas of Issuer"
                      },
                      "example": [
                        {
                          "$schema": "http://json-schema.org/draft-04/schema#",
                          "title": "BAMembershipType",
                          "description": "Membership Document for BA Credential",
                          "type": "object",
                          "properties": {
                            "firstName": {
                              "type": "string",
                              "pattern": "[a-zA-Z]+$",
                              "description": "First name of the user",
                              "maxLength": "0"
                            }
                          },
                          "required": [
                            "firstName"
                          ]
                        },
                        {
                          "$schema": "http://json-schema.org/draft-04/schema#",
                          "title": "GBAembershipType",
                          "description": "Membership Document for GBA Credential",
                          "type": "object",
                          "properties": {
                            "firstName": {
                              "type": "string",
                              "pattern": "[a-zA-Z]+$",
                              "description": "First name of the user",
                              "maxLength": "0"
                            },
                            "lastName": {
                              "type": "string",
                              "pattern": "[a-zA-Z]+$",
                              "description": "First name of the user",
                              "maxLength": "0"
                            }
                          },
                          "required": [
                            "firstName",
                            "lastName"
                          ]
                        }
                      ]
                    },
                    "message": {
                      "type": "string",
                      "example": "Schemas is retrieved"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Schema retrieval failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid Schemas"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}